fix(components): move research left panel to other file
