+++ src/lib/pages/account-details/components/resources/ResourceSectionBody.tsx
@@ -1,33 +1,14 @@
-import {
-  Accordion,
-  AccordionButton,
-  AccordionIcon,
-  AccordionItem,
-  AccordionPanel,
-  Badge,
-  Button,
-  Flex,
-  Heading,
-  Text,
-} from "@chakra-ui/react";
-import { truncate } from "lodash";
+import { Accordion, Badge, Button, Flex, Heading } from "@chakra-ui/react";
 import { useRouter } from "next/router";
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { useEffect, useState } from "react";
 
-import { ErrorFetching } from "../ErrorFetching";
-import { useInternalNavigate } from "lib/app-provider";
 import { CustomIcon } from "lib/components/icon";
-import InputWithIcon from "lib/components/InputWithIcon";
 import { Loading } from "lib/components/Loading";
-import { ResourceCard, ResourceDetailCard } from "lib/components/resource";
-import { EmptyState } from "lib/components/state";
-import type {
-  HumanAddr,
-  Option,
-  ResourceGroup,
-  ResourceGroupByAccount,
-} from "lib/types";
+import { ResourceDetailCard } from "lib/components/resource";
+import type { HumanAddr, Option, ResourceGroupByAccount } from "lib/types";
 import { getFirstQueryParam } from "lib/utils";
+
+import { ResourceLeftPanel } from "./ResourceLeftPanel";
 
 interface ResourceSectionBodyProps {
   address: HumanAddr;
@@ -41,36 +22,11 @@ export const ResourceSectionBody = ({
   isLoading,
 }: ResourceSectionBodyProps) => {
   const router = useRouter();
-  const navigate = useInternalNavigate();
-  const [keyword, setKeyword] = useState("");
+
   const [expandedIndexes, setExpandedIndexes] = useState<number[]>([]);
 
   const selectedAccountParam = getFirstQueryParam(router.query.account);
   const selectedNameParam = getFirstQueryParam(router.query.selected);
-
-  const handleSelectResource = useCallback(
-    (
-      account: ResourceGroupByAccount["owner"],
-      resource: ResourceGroup["group"]
-    ) => {
-      if (account === selectedAccountParam && resource === selectedNameParam)
-        return;
-      navigate({
-        pathname: `/accounts/[accountAddress]/[tab]`,
-        query: {
-          accountAddress: address,
-          tab: "resources",
-          account,
-          selected: resource,
-        },
-        replace: true,
-        options: {
-          shallow: true,
-        },
-      });
-    },
-    [selectedNameParam, selectedAccountParam, address, navigate]
-  );
 
   const updateExpandedIndexes = (indexes: number[]) =>
     setExpandedIndexes(indexes);
@@ -88,81 +44,19 @@ export const ResourceSectionBody = ({
     }
   }, [resourcesByOwner, selectedNameParam, selectedAccountParam]);
 
-  const filteredResourcesByOwner = useMemo(() => {
-    if (!keyword) return resourcesByOwner;
-
-    return resourcesByOwner?.map((each) => {
-      return {
-        ...each,
-        resources: each.resources.filter((resource) =>
-          resource.group
-            .toLowerCase()
-            .includes(keyword.trim().toLocaleLowerCase())
-        ),
-      };
-    });
-  }, [keyword, resourcesByOwner]);
-
   if (isLoading) return <Loading />;
-  if (!filteredResourcesByOwner) return <ErrorFetching />;
-  if (!filteredResourcesByOwner.length)
-    return <EmptyState imageVariant="empty" message="No resources found" />;
 
-  const selectedIndex = !selectedAccountParam
-    ? 0
-    : filteredResourcesByOwner.findIndex(
-        (item) => item.owner === selectedAccountParam
-      );
-  const selectedGroup = filteredResourcesByOwner[selectedIndex];
-  const selectedResources = selectedGroup.resources.find(
-    (item) => item.group === selectedNameParam
-  );
+  const selectedResource = resourcesByOwner
+    ?.find((resource) => resource.owner === selectedAccountParam)
+    ?.resources?.find((resource) => resource.group === selectedNameParam);
   return (
     <>
-      <Flex minW={{ base: "full", md: 80 }} direction="column">
-        <InputWithIcon
-          placeholder="Search with struct tag Name..."
-          value={keyword}
-          onChange={(e) => setKeyword(e.target.value)}
-          action="execute-message-search"
-        />
-        <Accordion
-          allowMultiple
-          defaultIndex={[selectedIndex]}
-          width="full"
-          mt={4}
-        >
-          {filteredResourcesByOwner.map((item) => (
-            <AccordionItem mb={4} key={item.owner}>
-              <AccordionButton>
-                <Flex p={4} justifyContent="space-between" w="full">
-                  <Text variant="body1" fontWeight={600}>
-                    {truncate(item.owner)}
-                  </Text>
-                  <AccordionIcon color="gray.600" />
-                </Flex>
-              </AccordionButton>
-              <AccordionPanel>
-                <Flex direction="column" gap={3} maxW={80}>
-                  {Object.values(item.resources).map((subitem) => (
-                    <ResourceCard
-                      key={subitem.group}
-                      name={subitem.group}
-                      amount={subitem.items.length}
-                      isSelected={selectedResources?.group === subitem.group}
-                      onClick={() =>
-                        handleSelectResource(item.owner, subitem.group)
-                      }
-                      hasBorder
-                    />
-                  ))}
-                </Flex>
-              </AccordionPanel>
-            </AccordionItem>
-          ))}
-        </Accordion>
-      </Flex>
-      {selectedResources && (
+      <ResourceLeftPanel
+        address={address}
+        resourcesByOwner={resourcesByOwner}
+        isLoading={isLoading}
+      />
+      {selectedResource && (
         <Flex direction="column" w="full">
           <Flex
             justifyContent="space-between"
@@ -172,15 +66,15 @@ export const ResourceSectionBody = ({
           >
             <Flex alignItems="center">
               <Heading as="h6" variant="h6" wordBreak="break-word">
-                {selectedGroup.owner}::{selectedResources.group}
+                {selectedResource.account}::{selectedResource.group}
               </Heading>
               <Badge variant="primary" ml={2}>
-                {selectedResources.items.length}
+                {selectedResource.items.length}
               </Badge>
             </Flex>
             <Button
               variant="outline-primary"
-              minW={28}
+              minW={32}
               size="sm"
               rightIcon={
                 <CustomIcon
@@ -191,7 +85,7 @@ export const ResourceSectionBody = ({
               onClick={() => {
                 setExpandedIndexes((prev) =>
                   !prev.length
-                    ? Array.from(Array(selectedResources.items.length).keys())
+                    ? Array.from(Array(selectedResource.items.length).keys())
                     : []
                 );
               }}
@@ -205,8 +99,11 @@ export const ResourceSectionBody = ({
             index={expandedIndexes}
             onChange={updateExpandedIndexes}
           >
-            {selectedResources.items.map((item) => (
-              <ResourceDetailCard resourceData={item} key={item.structTag} />
+            {selectedResource.items.map((item) => (
+              <ResourceDetailCard
+                resourceData={item}
+                key={`${item.structTag}`}
+              />
             ))}
           </Accordion>
         </Flex>
